splitted unsharing component
