+++ opencti-platform/opencti-graphql/src/modules/playbook/components/sharing-component.ts
